sale_pricelist_restrict_product_in_so - refactor one .py file per inherit !
