+++ sale_pricelist_restrict_product_in_so/__init__.py
@@ -20,4 +20,3 @@
 #
 ##############################################################################
 
-from . import product
