+++ sale_pricelist_restrict_product_in_so/__openerp__.py
@@ -1,5 +1,5 @@
 # -*- encoding: utf-8 -*-
-##############################################################################
+###############################################################################
 #
 #    OpenERP, Open Source Management Solution
 #    This module copyright (C) 2010 - 2014 Savoir-faire Linux
@@ -18,7 +18,7 @@
 #    You should have received a copy of the GNU Affero General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
-##############################################################################
+###############################################################################
 
 {
     'name': 'Price List Restrict Product in Sales',
@@ -38,6 +38,7 @@ allowed by the current price list.
 
 Contributors
 ------------
+* Marc Cassuto (marc.cassuto@savoirfairelinux.com)
 * Vincent Vinet (vincent.vinet@savoirfairelinux.com)
 """,
     'depends': [
